# .history/docs/CASHFLOW_MANAGEMENT_SERVICE_OPENAPI_20250828185140.yaml
openapi: 3.0.3
info:
  title: Cash Flow Management Service API
  description: |
    Comprehensive API for managing equity swap cashflows throughout their lifecycle.
    
    ## Overview
    This service handles the complete lifecycle of cashflows including:
    - **Interest & Dividend Accruals**: Daily accrual calculations with business day adjustments
    - **Performance Cashflows**: Unrealized P&L calculations for equity returns
    - **State Management**: Accrued → Realized Deferred → Realized Unsettled → Settled
    - **Deferral Management**: Business rule-based cashflow deferrals
    - **Reporting & Analytics**: Comprehensive cashflow reporting and analysis
    
    ## Technology Stack
    - **Java 21**: Latest LTS with virtual threads, pattern matching, and performance improvements
    - **Spring Boot 3.3+**: Latest version with Java 21 support and reactive programming
    - **Virtual Threads**: High-throughput concurrent processing (100,000+ concurrent operations)
    - **Thread Partitioning**: Contract + Underlier (securityId) + Calculation Type based isolation
    - **Project Reactor**: Reactive streams for high-volume processing with backpressure handling
    - **Spring WebFlux**: Non-blocking reactive web framework
    - **Pattern Matching**: Clean, readable code for complex cashflow type handling
    - **Sealed Classes**: Type-safe cashflow status and type hierarchies
    
    ## Cashflow States
    1. **ACCRUED**: Initial state with daily accrual tracking
...
    
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /cashflows/generate:
    post:
      summary: Generate Cashflows
      description: Generate cashflows for specified contracts and calculation dates
      tags: [Cashflow Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashflowGenerationRequest'
      responses:
        '202':
          description: Cashflow generation accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowGenerationResponse'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Business rule validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /cashflows/generate/interest:
    post:
      summary: Generate Interest Cashflows
      description: Generate interest accrual cashflows for specified contracts
      tags: [Cashflow Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InterestCashflowGenerationRequest'
      responses:
        '202':
          description: Interest cashflow generation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowGenerationResponse'
  /cashflows/generate/dividend:
    post:
      summary: Generate Dividend Cashflows
      description: Generate dividend cashflows based on corporate actions
      tags: [Cashflow Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DividendCashflowGenerationRequest'
      responses:
        '202':
          description: Dividend cashflow generation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowGenerationResponse'
  /cashflows/generate/performance:
    post:
      summary: Generate Performance Cashflows
      description: Generate unrealized P&L cashflows for equity performance
      tags: [Cashflow Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerformanceCashflowGenerationRequest'
      responses:
        '202':
          description: Performance cashflow generation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowGenerationResponse'

  # Reactive Streaming Endpoints
  /cashflows/generate/reactive:
    post:
      summary: Generate Cashflows Reactively
      description: Generate cashflows using reactive streams with backpressure handling
      tags: [Reactive Processing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashflowGenerationRequest'
      responses:
        '200':
          description: Cashflows generated reactively with streaming response
          content:
            application/stream+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cashflow'
  /cashflows/generate/stream:
    post:
      summary: Generate Cashflows with Server-Sent Events
      description: Generate cashflows using Server-Sent Events for real-time streaming
      tags: [Reactive Processing]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CashflowGenerationRequest'
      responses:
        '200':
          description: Server-Sent Events stream of cashflow generation
          content:
            text/event-stream:
              schema:
                $ref: '#/components/schemas/ServerSentEvent'

  # Calculation Type Specific Endpoints
...
  /cashflows:
    get:
      summary: Search Cashflows
      description: Search and filter cashflows with pagination
      tags: [Cashflow Queries]
      parameters:
        - name: contractId
          in: query
          schema:
            type: string
            format: uuid
        - name: legId
          in: query
          schema:
            type: string
            format: uuid
        - name: cashflowType
          in: query
          schema:
            $ref: '#/components/schemas/CashflowType'
        - name: cashflowStatus
          in: query
          schema:
            $ref: '#/components/schemas/CashflowStatus'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: currency
          in: query
          schema:
            type: string
            minLength: 3
            maxLength: 3
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 100
            maximum: 1000
        - name: sort
          in: query
          schema:
            type: string
            default: "createdAt,desc"
      responses:
        '200':
          description: Cashflows retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashflowPageResponse'
  /cashflows/{cashflowId}:
    get:
      summary: Get Cashflow by ID
      description: Retrieve a specific cashflow by its ID
      tags: [Cashflow Queries]
      parameters:
        - name: cashflowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Cashflow retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cashflow'
        '404':
          description: Cashflow not found

  # Unrealized P&L Management
...
  /threads/partitions/{partitionKey}:
    get:
      summary: Get Specific Partition Status
      description: Retrieve status of a specific thread partition
      tags: [Thread Management]
      parameters:
        - name: partitionKey
          in: path
          required: true
          schema:
            type: string
          description: Partition key in format contractId:securityId:calculationType
      responses:
        '200':
          description: Partition status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartitionDetail'
  /threads/partitions/{partitionKey}/operations:
    get:
      summary: Get Partition Operations
      description: Retrieve current operations running in a specific partition
      tags: [Thread Management]
      parameters:
        - name: partitionKey
          in: path
          required: true
          schema:
            type: string
        - name: includeHistory
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Partition operations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartitionOperations'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    CashflowType:
      type: string
      enum: [INTEREST, DIVIDEND, PERFORMANCE, FEES]
      description: Type of cashflow
    CashflowStatus:
      type: string
      enum: [ACCRUED, REALIZED_DEFERRED, REALIZED_UNSETTLED, REALIZED_SETTLED]
      description: Current status of the cashflow
    CashflowSubType:
      type: string
      enum: [PRICE_RETURN, DIVIDEND_RETURN, INTEREST_ACCRUAL, MANAGEMENT_FEE, PERFORMANCE_FEE]
      description: Subtype of the cashflow
    DeferralReason:
      type: string
      enum: [BUSINESS_RULE, REGULATORY, COUNTERPARTY_REQUEST, MARKET_CONDITIONS, TECHNICAL_ISSUE]
      description: Reason for cashflow deferral
    CalculationMethod:
      type: string
      enum: [MARK_TO_MARKET, MODEL_VALUATION, EXTERNAL_PRICE, INTERNAL_MODEL]
      description: Method used for cashflow calculation
...
    CashflowGenerationRequest:
      type: object
      required: [contractIds, calculationDate, cashflowTypes]
      properties:
        contractIds:
          type: array
          items:
            type: string
            format: uuid
          description: List of contract IDs to generate cashflows for
        legIds:
          type: array
          items:
            type: string
            format: uuid
          description: Specific leg IDs (optional, generates for all legs if not specified)
        calculationDate:
          type: string
          format: date
          description: Date for cashflow calculation
        cashflowTypes:
          type: array
          items:
            $ref: '#/components/schemas/CashflowType'
          description: Types of cashflows to generate
        marketDataOverride:
          $ref: '#/components/schemas/MarketDataOverride'
        businessRules:
          $ref: '#/components/schemas/BusinessRules'
    CashflowGenerationResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          description: ID of the background job processing the request
        estimatedCompletionTime:
          type: string
          format: date-time
          description: Estimated time when processing will complete
        contractsProcessed:
          type: integer
          description: Number of contracts being processed
        message:
          type: string
          description: Human-readable message about the request
    InterestCashflowGenerationRequest:
      type: object
      required: [contractIds, calculationDate, interestRate]
      properties:
        contractIds:
          type: array
          items:
            type: string
            format: uuid
        calculationDate:
          type: string
          format: date
        interestRate:
          type: number
          format: decimal
          description: Interest rate to use for calculation
        dayCountConvention:
          type: string
          enum: [ACT_365, ACT_360, ACT_ACT, 30_360]
          default: ACT_365
        businessDayAdjustment:
          type: string
          enum: [FOLLOWING, PRECEDING, MODIFIED_FOLLOWING]
          default: FOLLOWING
    DividendCashflowGenerationRequest:
      type: object
      required: [contractIds, dividendDate, dividendAmount]
      properties:
        contractIds:
          type: array
          items:
            type: string
            format: uuid
        dividendDate:
          type: string
          format: date
        dividendAmount:
          type: number
          format: decimal
        dividendType:
          type: string
          enum: [ORDINARY, SPECIAL, RETURN_OF_CAPITAL]
          default: ORDINARY
        exDate:
          type: string
          format: date
        recordDate:
          type: string
          format: date
    PerformanceCashflowGenerationRequest:
      type: object
      required: [contractIds, valuationDate]
      properties:
        contractIds:
          type: array
          items:
            type: string
            format: uuid
        valuationDate:
          type: string
          format: date
        marketData:
          $ref: '#/components/schemas/MarketData'
        calculationMethod:
          $ref: '#/components/schemas/CalculationMethod'
        includeComponents:
          type: boolean
          default: true
          description: Whether to include P&L component breakdown
...
    Cashflow:
      type: object
      required: [id, contractId, legId, cashflowType, amount, currency, cashflowStatus]
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        legId:
          type: string
          format: uuid
        performancePayoutId:
          type: string
          format: uuid
        cashflowReference:
          type: string
        cashflowType:
          $ref: '#/components/schemas/CashflowType'
        cashflowSubType:
          $ref: '#/components/schemas/CashflowSubType'
        amount:
          type: number
          format: decimal
        currency:
          type: string
          minLength: 3
          maxLength: 3
        fxRate:
          type: number
          format: decimal
          default: 1.0
        baseCurrencyAmount:
          type: number
          format: decimal
        payerPartyId:
          type: string
          format: uuid
        receiverPartyId:
          type: string
          format: uuid
        accrualStartDate:
          type: string
          format: date
        accrualEndDate:
          type: string
          format: date
        calculationDate:
          type: string
          format: date
        paymentDate:
          type: string
          format: date
        cashflowStatus:
          $ref: '#/components/schemas/CashflowStatus'
        realizationDate:
          type: string
          format: date
        deferralDate:
          type: string
          format: date
        settlementDate:
          type: string
          format: date
        deferralReason:
          $ref: '#/components/schemas/DeferralReason'
        deferralPeriodDays:
          type: integer
        expectedRealizationDate:
          type: string
          format: date
        calculationMethod:
          $ref: '#/components/schemas/CalculationMethod'
        calculationSource:
          type: string
        marketDataSnapshot:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        businessMetadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
...
    MarketData:
      type: object
      properties:
        prices:
          type: object
          additionalProperties:
            type: number
            format: decimal
        interestRates:
          type: object
          additionalProperties:
            type: number
            format: decimal
        fxRates:
          type: object
          additionalProperties:
            type: number
            format: decimal
        dividendData:
          type: object
          additionalProperties:
            type: object
    MarketDataOverride:
      type: object
      properties:
        prices:
          type: object
          additionalProperties:
            type: number
            format: decimal
        rates:
          type: object
          additionalProperties:
            type: number
            format: decimal
        effectiveDate:
          type: string
          format: date-time
    BusinessRules:
      type: object
      properties:
        deferralRules:
          type: array
          items:
            type: object
        calculationRules:
          type: object
        validationRules:
          type: object

    # Response Models
    CashflowPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Cashflow'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
        totalPages:
          type: integer
        last:
          type: boolean
        first:
          type: boolean
        size:
          type: integer
        number:
          type: integer
...
    Pageable:
      type: object
      properties:
        pageNumber:
          type: integer
        pageSize:
          type: integer
        sort:
          type: array
          items:
            type: object
            properties:
              property:
                type: string
              direction:
                type: string
                enum: [ASC, DESC]

    # Error Models
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        details:
          type: array
          items:
            type: object
    ValidationErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        validationErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              rejectedValue:
                type: object

    # Batch Operation Models
...